+++ tests/ui/specs/TimePicker/TimePicker-specs.js
@@ -16,19 +16,19 @@ describe('TimePicker', function () {
 			});
 
 			it('should have hour-minute-meridiem order', function () {
-				browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  timeoutMsg: 'hour focused'});
+				browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500, timeoutMsg: 'hour focused'});
 				Page.spotlightRight();
-				browser.waitUntil(() => timePicker.minute.isFocused(), {timeout: 1500,  timeoutMsg: 'minute focused'});
+				browser.waitUntil(() => timePicker.incrementer(timePicker.minute).isFocused(), {timeout: 1500, timeoutMsg: 'minute focused'});
 				Page.spotlightRight();
-				browser.waitUntil(() => timePicker.meridiem.isFocused(), {timeout: 1500,  timeoutMsg: 'meridiem focused'});
+				browser.waitUntil(() => timePicker.incrementer(timePicker.meridiem).isFocused(), {timeout: 1500, timeoutMsg: 'meridiem focused'});
 			});
 
 			describe('5-way', function () {
 				// Start of [QWT-2546] - Hour, Minute, Meridiem pickers Animates with 5-way - LTR
 				it('should increase the hour when incrementing the picker', function () {
 					const {hour} = extractValues(timePicker);
-					browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  interval: 100});
-					Page.spotlightUp();
+					browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500,  interval: 100});
+					Page.spotlightSelect();
 					const {hour: value} = extractValues(timePicker);
 					const expected = hour < 12 ? hour + 1 : 1;
 					expect(value).to.equal(expected);
@@ -36,8 +36,9 @@ describe('TimePicker', function () {
 
 				it('should decrease the hour when decrementing the picker', function () {
 					const {hour} = extractValues(timePicker);
-					browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  interval: 100});
+					browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500,  interval: 100});
 					Page.spotlightDown();
+					Page.spotlightSelect();
 					const {hour: value} = extractValues(timePicker);
 					const expected = hour > 1 ? hour - 1 : 12;
 					expect(value).to.equal(expected);
@@ -46,8 +47,8 @@ describe('TimePicker', function () {
 				it('should increase the minute when incrementing the picker', function () {
 					const {minute} = extractValues(timePicker);
 					Page.spotlightRight();
-					browser.waitUntil(() => timePicker.minute.isFocused(), {timeout: 1500,  interval: 100});
-					Page.spotlightUp();
+					browser.waitUntil(() => timePicker.incrementer(timePicker.minute).isFocused(), {timeout: 1500,  interval: 100});
+					Page.spotlightSelect();
 					const {minute: value} = extractValues(timePicker);
 					const expected = minute !== 59 ? minute + 1 : 0;
 					expect(value).to.equal(expected);
@@ -56,8 +57,9 @@ describe('TimePicker', function () {
 				it('should decrease the minute when decrementing the picker', function () {
 					const {minute} = extractValues(timePicker);
 					Page.spotlightRight();
-					browser.waitUntil(() => timePicker.minute.isFocused(), {timeout: 1500,  interval: 100});
+					browser.waitUntil(() => timePicker.incrementer(timePicker.minute).isFocused(), {timeout: 1500,  interval: 100});
 					Page.spotlightDown();
+					Page.spotlightSelect();
 					const {minute: value} = extractValues(timePicker);
 					const expected = minute !== 0 ? minute - 1 : 59;
 					expect(value).to.equal(expected);
@@ -67,8 +69,8 @@ describe('TimePicker', function () {
 					const time = timePicker.timeLabel;
 					Page.spotlightRight();
 					Page.spotlightRight();
-					browser.waitUntil(() => timePicker.meridiem.isFocused(), {timeout: 1500,  interval: 100});
-					Page.spotlightUp();
+					browser.waitUntil(() => timePicker.incrementer(timePicker.meridiem).isFocused(), {timeout: 1500,  interval: 100});
+					Page.spotlightSelect();
 					const newTime = timePicker.timeLabel;
 					const value = time !== newTime;
 					expect(value).to.equal(true);
@@ -78,8 +80,9 @@ describe('TimePicker', function () {
 					const time = timePicker.timeLabel;
 					Page.spotlightRight();
 					Page.spotlightRight();
-					browser.waitUntil(() => timePicker.meridiem.isFocused(), {timeout: 1500,  interval: 100});
+					browser.waitUntil(() => timePicker.incrementer(timePicker.meridiem).isFocused(), {timeout: 1500,  interval: 100});
 					Page.spotlightDown();
+					Page.spotlightSelect();
 					const newTime = timePicker.timeLabel;
 					const value = time !== newTime;
 					expect(value).to.equal(true);
@@ -88,9 +91,10 @@ describe('TimePicker', function () {
 
 				it('should change the meridiem on hour boundaries', function () {
 					const {meridiem} = extractValues(timePicker);
+					Page.spotlightDown();
 					// 12 hours ought to change the value text if meridiem changes
 					for (let i = 12; i; i -= 1) {
-						Page.spotlightDown();
+						Page.spotlightSelect();
 					}
 					expect(meridiem !== timePicker.item(timePicker.meridiem).getText()).to.be.true();
 				});
@@ -99,8 +103,8 @@ describe('TimePicker', function () {
 			describe('pointer', function () {
 
 				it('should select hour when opened', function () {
-					timePicker.hour.click();
-					browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  interval: 100});
+					timePicker.incrementer(timePicker.hour).click();
+					browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500,  interval: 100});
 				});
 
 				// Start of [QWT-2552] - Hour, Minute, Meridiem pickers Animate on Pointer Click and Hold
@@ -122,8 +126,6 @@ describe('TimePicker', function () {
 
 				it('should increase the minute when incrementing the picker]', function () {
 					const {minute} = extractValues(timePicker);
-					timePicker.minute.click();
-					browser.waitUntil(() => timePicker.minute.isFocused(), {timeout: 1500,  interval: 100});
 					timePicker.incrementer(timePicker.minute).click();
 					const {minute: value} = extractValues(timePicker);
 					const expected = minute !== 59 ? minute + 1 : 0;
@@ -132,8 +134,6 @@ describe('TimePicker', function () {
 
 				it('should decrease the minute when decrementing the picker', function () {
 					const {minute} = extractValues(timePicker);
-					timePicker.minute.click();
-					browser.waitUntil(() => timePicker.minute.isFocused(), {timeout: 1500,  interval: 100});
 					timePicker.decrementer(timePicker.minute).click();
 					const {minute: value} = extractValues(timePicker);
 					const expected = minute !== 0 ? minute - 1 : 59;
@@ -182,9 +182,6 @@ describe('TimePicker', function () {
 
 			describe('pointer', function () {
 				it('should not update hour on click', function () {
-					timePicker.hour.click();
-					expect(timePicker.hour.isFocused()).to.be.true();
-
 					timePicker.incrementer(timePicker.hour).click();
 					browser.pause(500);
 					const {hour: value} = extractValues(timePicker);
@@ -198,16 +195,10 @@ describe('TimePicker', function () {
 			it('should not update \'defaultValue\' on click', function () {
 				const {hour, minute, meridiem} = extractValues(timePicker);
 
-				timePicker.minute.click();
-				expect(timePicker.minute.isFocused()).to.be.true();
 				timePicker.decrementer(timePicker.minute).click();
 
-				timePicker.hour.click();
-				expect(timePicker.hour.isFocused()).to.be.true();
 				timePicker.decrementer(timePicker.hour).click();
 
-				timePicker.meridiem.click();
-				expect(timePicker.meridiem.isFocused()).to.be.true();
 				timePicker.decrementer(timePicker.meridiem).click();
 
 				browser.pause(500);
@@ -227,17 +218,17 @@ describe('TimePicker', function () {
 		});
 
 		it('should focus middle picker (hour) when selected', function () {
-			browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  interval: 100});
+			browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500, interval: 100});
 		});
 
 		it('should have minute-hour-meridiem order', function () {
-			browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  timeoutMsg: 'initial', interval: 100});
+			browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500, timeoutMsg: 'initial', interval: 100});
 			Page.spotlightRight();
-			browser.waitUntil(() => timePicker.minute.isFocused(), {timeout: 1500,  timeoutMsg: 'minute', interval: 100});
+			browser.waitUntil(() => timePicker.incrementer(timePicker.minute).isFocused(), {timeout: 1500, timeoutMsg: 'minute', interval: 100});
 			Page.spotlightLeft();
-			browser.waitUntil(() => timePicker.hour.isFocused(), {timeout: 1500,  timeoutMsg: 'hour', interval: 100});
+			browser.waitUntil(() => timePicker.incrementer(timePicker.hour).isFocused(), {timeout: 1500, timeoutMsg: 'hour', interval: 100});
 			Page.spotlightLeft();
-			browser.waitUntil(() => timePicker.meridiem.isFocused(), {timeout: 1500,  timeoutMsg: 'meridiem', interval: 100});
+			browser.waitUntil(() => timePicker.incrementer(timePicker.meridiem).isFocused(), {timeout: 1500, timeoutMsg: 'meridiem', interval: 100});
 		});
 	});
 
@@ -257,13 +248,6 @@ describe('TimePicker', function () {
 		});
 
 		it('should increment hours from 23 to 0', function () {
-			// FIXME: replace with trigger for expandable open (or replace with defaultOpen picker)
-			Page.delay(500);
-
-			// go to 23 first
-			timePicker.hour.click();
-			expect(timePicker.hour.isFocused()).to.be.true();
-
 			timePicker.decrementer(timePicker.hour).click();
 			expect(extractValues(timePicker).hour).to.equal(23);
 			// now increment
@@ -272,11 +256,6 @@ describe('TimePicker', function () {
 		});
 
 		it('should decrement hours from 0 to 23', function () {
-			// FIXME: replace with trigger for expandable open (or replace with defaultOpen picker)
-			Page.delay(500);
-			timePicker.hour.click();
-			expect(timePicker.hour.isFocused()).to.be.true();
-
 			timePicker.decrementer(timePicker.hour).click();
 			expect(extractValues(timePicker).hour).to.equal(23);
 		});