+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Changed
 
-- `sandstone/DatePicker` and `sandstone/TimePicker` to not show press effect via touch input
+- `sandstone/DatePicker` and `sandstone/TimePicker` to not joined picker
 - `sandstone/RadioItem` to use `onClick` instead of `onTap` for touch support
 - `sandstone/Scroller` and `sandstone/VirtualList` to show overscroll effect when flicking